requirements.txt
pyTelegramBotAPI==4.12.0
requests==2.31.0
python-dotenv==0.21.1